refactor: update validators with improved error handling
